[TASK] Splits GitHub rendering action into composite (#169)
This PR splits the current process of single-use rendering up into a
two-step composite process as outlined in
https://github.com/TYPO3-Documentation/render-guides/issues/158

The first step takes care of configuring a local guides.xml (during
rendering/GH-action only) with dynamic parameters. Those parameters
fixate the input from the GitHub action for the target GIT branch and
target deployment directory.

The second step is the usual rendering, which will then take into
account the modified parameters, so that the rendered output includes
also the non-committed changes, as the old sphinx jobfile did.

The commit includes a large addition to the `ConfigureCommand` that is
needed to configure the required attributes. That code will need later
refactoring and added unit testing. For now it should work to progress
on the GitHub action.

This PR should be merged in conjunction with the changes made to the
`gh-render-action` PR:
https://github.com/TYPO3-Documentation/gh-render-action/pull/21