[TASK] Streamline documentation for subtree-split repositories (#395)
I noticed that the README.rst files for our subtree repositories are
shown in the GitHub repositories, and do not properly indicate that
those are subtree repos.

I tried to unify all four package READMEs, rather quick and dirty.
Please check if this is suitable. Feel free to add commits to this
branch to change wordings.